Create README.md
In this project, the logic of a 12-bit asynchronous up-down counter with parallel load is developed and verified using class-based testbench. Synthesis and physical design of the design is also carried out. A 12-bit counter can count from 0 to 2^12 – 1, or 0 to 4095. It increments or decrements the count value in every clock pulse. A reset button sets the count value to zero (0), and starts the counter again. An up-down counter can count do both up-counting and down-counting operations. But, this requires an additional input for the counter to determine whether it will do up-count or down-count operation. That input is denoted by up_down. At any moment, the counter can start counting from a specific value, denoted by data, and data value will be loaded to count when input load is set to 1. The counter works normally when enable is 1, and stops counting or loading data if enable is 0. All the inputs except up_down are asynchronous.